fix(dataTable): refactor datatable to v2 #closes1662
