+++ packages/geoview-core/src/core/components/data-table/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useLightbox';
-export * from './useSelectedRowMessage';
-export * from './useFilteredRowMessage';
+export * from './useSelectedRows';
+export * from './useFilteredRows';
+export * from './useToolbarActionMessage';
