+++ packages/geoview-core/src/core/components/index.ts
@@ -11,7 +11,7 @@ export * from './data-table/data-table-api';
 export * from './data-table/export-button';
 export * from './data-table/filter-map';
 export * from './data-table/json-export-button';
-export * from './data-table/map-data-table';
+export * from './data-table/data-table';
 export * from './details/details-panel';
 export * from './export/export-modal-button';
 export * from './export/export-modal';
